add .gitignore

+++ chart/candlestick/candlestick.go
@@ -16,7 +16,7 @@ type Candle struct {
   High    float64
   Low    float64
   Close    float64
-  Volume    float64
+  Volume    int64
 }
 
 type Candlestick struct {
